Basically a complete rewrite of automatically removing expired cards.  Combine this with a new cron task of removing old unused cards.  Keep track of when a card is used (added or charged) and remove cards that haven't been used in over a year.  Update indexes as well to clean them up and remove old ones.

+++ pkgs/timestamps/timestamps.go
@@ -10,10 +10,10 @@ package timestamps
 import "time"
 
 //Unix returns the number of seconds as an integer since epoch
-func Unix() int {
+func Unix() int64 {
 	t := time.Now()
 	s := t.Unix()
-	return int(s)
+	return s
 }
 
 //ISO8601 returns a datetime with the format YY-MM-DDTHH:MM:SS.mmmZ
